+++ components/VideoContainer.tsx
@@ -5,8 +5,8 @@ const VideoContainer = ({ title, videoId }) => {
   return (
     <div className="w-full h-[550px] flex justify-center items-center my-10 bg-blackX relative">
       <Player videoId={videoId} />
-      <div className="flex absolute bottom-4">
-        <h2 className="p-2 text-center uppercase text-blackX sm:text-5xl text-3xl font-bold transition-all ease-in bg-whiteX bg-opacity-30">
+      <div className="flex absolute bottom-0 transition-all ease-in-out duration-150 hover:hidden">
+        <h2 className="z-10 text-white font-bold sm:text-5xl text-3xl p-1 uppercase bg-blackX bg-opacity-60 w-full text-center">
           {title}
         </h2>
       </div>